index内でfindContactPageを定義
